Angler: Create a link stm_sensor
QC write a common HAL to access laser device. It is not only used in
Angler. So we need to supply a common device name on it. In addition,
laser device is registed as input device, so we need search all the
input device and find the matched device whose name is
"STM VL6180 proximity sensor".

BUG=22281124

Change-Id: If63fc81b74e31ca9c1e52a8a1d45b127787039ee