Camera: LDAF: fix camera access calbration issue
Laser calibration data are stored under /persist/data, it's access
right is 0x0700, so camera have no right to access it. We need to
change it to 0x0710 to ensure system group have right to access it.
It will not affect other file or directory under it, they have it's
own access right.

BUG: 22772098
Change-Id: I06009f7d95bd9a75a815d23f1042b1e7504e0ce2