Fixed the wrong carrier name displayed on the UI
The spn from EF_SPN on LGE U+ SIM is different
from the plmn. Thus the UI shows "LG U+|LGU+"
as the carrier name on the UI. The fix is actually
a workaround to override the spn with "LG U+" which is
same as plmn name. The duplicate detection mechanism
will automatically rip off the spn from the name
displayed on the UI.

bug: 23506080
Change-Id: I83b6b2dce6d06b4382b293c2e6e8a5d4b8ac3e8c