Allow alternative sensorhub implementation
Nanohub requires the use of an alternative Linux kernel,
sensor HALs, and related utils. When
TARGET_USES_CHINOOK_SENSORHUB is false or unset, use
the alternatives. The change also moves the
TARGET_USES_CHINOOK_SENSORHUB from vendor to device.
The variable is used in both places so the location
should be in device, not vendor (since vendor is
optional).

Change-Id: I61c3efbbc022946dd3fad8c05857264548d3113d