dontaudit perfd self:capability kill;
perfd sends a kill(PID, 0) (a "null" kill) to determine if a process
exists. A process exists if the kill succeeds, or it returns with
errno=EPERM. A process doesn't exist if the kill fails with
errno=ESRCH. In either case, CAP_KILL is not needed to get this
information. Suppress the SELinux denials.

Change-Id: I177d432d96de938be6acf9076352d9df18e6308a