Update patch set 2
Patch Set 2:

This is nice, but i have some concerns.

1. It initially didn't work for me on debian. I figured out I was missing simg2img util, but there was no message
saying it was missing. I had to install android-tools-fsutil package to get that util. That package is not normally
needed for android building so I figure most people will not have it. But not sure if it's included in other linux
distros.

2. This deveates from the normal way cm does it on every other device, so I imagine it will cause a bit of confusion.

3. The blobs in a factory image aren't always exactly whats in cm.

My concerns might be very minor, but just wanted to bring them up to get opinions and make sure everything is taken
into consideration.

Patch-set: 2