Update patch set 6
Patch Set 6:

the thing i dont understand is why we have to use a prebuilt camera hal. usually nexus devices use the aosp sources... writing a wrapper seems over complicated for this type of device but its the only way I know how to do it? whos the maintainer of this device?

Patch-set: 6