Update patch set 3
Patch Set 3:

> Just a shot in the dark - could the issue be with the updated kati
 > from google/kati?

could be but feels more like a compiler or art issue.  for a given build, it's the same bad pointer address across multiple binaries and essentially the same stack trace leading up to it.

why does the same dex2oat code work fine on arm64 but seemingly produce incorrect output on an x86 host ?...  (As a side note, I see that art aosp/master is way ahead of android-7.0.0_r6.)

Patch-set: 3