Update patch set 1
Patch Set 1:

> razorloves, I think you mentioned in the past how fi needs other
 > things too.  Do we know if they have to matched to the version of
 > tycho ?  Just thinking about the value of including it.
 > 
 > The rest LGTM (unable to test because I'm away from home).

All that is needed for fi appears to be included here. ciwrl tested it. But i dont recall any issues with matching versions.
Since not all gapps zips out there include all the fi stuff, i think it's best to add to rom.  And even though some of the apk's are available in the playstore, some of them need to be installed to /system, so that's another reason.

Patch-set: 1