Bump some angler performance expectations
We've been seeing some angler CTS failures where a couple codecs were
running faster than projected in media_codecs_performance.xml (yes,
it is currently a bug to be too fast). This raises the upper bounds
for the couple of tests that were exceeding these speeds.

An independent effort is trying to automate the work of mining test data
for achieved speeds, doing the appropriate statistical crunching, and
generating a new file automatically.  That isn't ready at this time.

Bug: 62359539
Test: successful rerun of CTS for previously failing test cases