Update patch set 1
Patch Set 1:

This bug is also in the stock rom. My hunch is that google has it configured for fingerprint sensors on the front of the device.
This bug can be corrected in a couple different places. I consider this a temporary fix until google fixes it properly. Most likely in the fpc firmware or driver in the kernel.

Patch-set: 1
Label: Verified=0