Update patch set 3
Patch Set 3:

> We currently don't need the "build nfc from source" change since
 > the 15.0 branch doesn't have this change (yet).
 > https://github.com/LineageOS/android_external_libnfc-nci/commit/d0b9d073a2f878910c7a8ed77f1859e931feb589

Thanks, ACK.

I was thinking that we would wait until it's possible to build without needing any picks.  In which case, perhaps we leave as it for now and we review again later (because who knows how long it will be before we can build without picks).

Patch-set: 3
Label: Verified=0