Revert "Angler: Remove unnecessary WITH_DEXPREOPT block"
This reverts commit 150431c76a9f2637e55b4296cc3232bc2ff80685.

DEXPREOPT is not enabled by default in lineage, like it is in AOSP.
This revert brings back the WITH_DEXPREOPT block and enables it.

Note: We disabled DEXPREOPT on angler in nougat due to crashes.
No signs of that problem in oreo, so lets get those fast initial boots back.

Change-Id: Ia021473311f942e0387e7b16b005f3e2429a8b08