Update patch set 1
Patch Set 1:

Today i got the chance to test. Dirty flash with a build with the two commits: no issues.
Clean flash: no SIM anymore.
Here are the two logs:
https://drive.google.com/open?id=0Bzua1MirSibkREJCcUd6Uk5nanc

PS: i'm on a personal aosp based build so it could be only my issue, in that case sorry and nevermind.

Patch-set: 1