Update patch set 1
Patch Set 1:

Just for experimentation purposes.

It boots but, so far, I've noticed at least the following issues:
*) On dirty data, systemui navbar home and recents buttons don't work.  quicksettings pulldown doesn't work properly and you can't enter settings except via the launcher.  settings things mobile data is always off (even when it's actually working).  These issues don't matter if they don't occur during cm-14.1 -> r23 lineage-15.0 upgrade.

*) On clean data, sim card is not detected but navbar and quicksettings issues are no longer present.

I think gps might not have been working but didn't test properly.

Some of these issues will be because we're tracking a pixel2 release tag (r23) but some might be because of the vendor build by this gerrit.

Patch-set: 1